PGEN-RGX-0078-0212 (leaf RGX-0078.5.j.4): RESULT-CARRIER SLIMMING LANDED (the NINTH -0197 ratchet fix) - ParseNode 72->48B (thin &'static &'static str rule-name + thin Quantified label + Copy Span{u32,u32} + honest 4GiB parse-entry refusal); pooled-paired corpus geomean 1018.404 -> 994.364 ns = -2.3606% (flips 0/2189; MAX 372,542 <= 425,000; EVERY band improves); floor -> 1,004.4 ns - the FIRST sub-1us raw corpus geomean readings
- Design-prereg-first: thin-ref ADJUDICATED over the u16-id table (equal 48B by alignment,
  zero table/serde machinery; serde &T-delegation keeps the wire byte-identical); u8 kind
  REFUTED (open kind set "0,127"); compile-time const size pins 48/32/8.
- Scope: lib core (mod.rs) + ~160 mechanical consumer sites + all live emitter templates
  (ast_based_generator + cascade + cascade/value + scan + the string-template emitters) +
  interpreter intern_ref; PgenValue/store/memo/tape/parse decisions untouched.
- Migration: the -0211-banked cold-bootstrap tax executed by the proven recipe (seed ->
  annotation pair -> wave -> focus x8 -> dual-feature LAST, peak 10,081MB guarded); ebnf
  FIXED POINT byte-identical under the final all-11 binary; artifacts regex 1b5bdcf0 /
  ebnf d4c257cf (all 11 banked pre/post).
- Correctness: explicit wire-format oracle 44/44 base-vs-cand typed-AST dumps
  BYTE-IDENTICAL (+ identical reject sets); battery green: dual-feature lib 1006/0/29
  incl. the ALL-11 oracle; cert x3 byte-exact 268/9/259/0 fully_certified; shape/duality/
  PCRE2-oracle; clippy source-strict.
- Measurement custody: run 1 INVALIDATED (foreign harness test binary at 101% CPU; base
  +15.62% vs its own -0210 sweep - tool-pinned live); hardened interleaved rerun protocol
  PRE-REGISTERED before numbers (094a3ad7) and adjudicated custody-green (base -0.99%/
  -0.92% vs floor; base1<->base2 -0.08%); THIRD consecutive fix INSIDE its priced band
  (predicted -1.0..-3.4%).
- Floor of record: corpus 1,004.4 ns single-sweep-comparable / bench ~1,579.7 ns ~314x;
  cumulative nine fixes 1,263.4 -> 1,004.4 ~ -20.5%; margined <=950.0 bar needs -5.4%.
- Lockstep: contract 1.1.108 (Rust-level carrier field types + the 4GiB honest bound; wire
  schema unchanged 1) + regex book public-api envelope + speed-journey/SUMMARY/mechanics
  scoreboards + LIVE tracker note + MEMORY/TASK_TREE/CHANGES/DEV_NOTES; candidate probe
  preserved regex_perf_probe_carrier48_8d392176. Session STOPS after this clean commit;
  NEXT (fresh session) = re-price on the new floor.

+++ rust/src/ast_pipeline/ast_based_generator/cascade/value.rs
@@ -861,7 +861,7 @@ impl AstBasedGenerator {
                             Ok(quote! {
                                 {
                                     let __pgen_alt_node = parser.deriv_next_boundary();
-                                    parser.deriv_pos = __pgen_alt_node.span.end;
+                                    parser.deriv_pos = __pgen_alt_node.span.end as usize;
                                     __pgen_alt_node.content.to_shaped_value(parser.arena)
                                 }
                             })
@@ -1096,7 +1096,7 @@ impl AstBasedGenerator {
                             Ok(quote! {
                                 {
                                     let __pgen_alt_node = parser.deriv_next_boundary();
-                                    parser.deriv_pos = __pgen_alt_node.span.end;
+                                    parser.deriv_pos = __pgen_alt_node.span.end as usize;
                                 }
                             })
                         }
